Move footer and header into components directory
