+++ app/(site)/components/footer.module.css
@@ -1,30 +1,3 @@
-.nav {
-  display: flex;
-  justify-content: space-between;
-  align-items: stretch;
-  border-bottom: 2px solid var(--foreground);
-}
-
-p.homeLink {
-  font-family: var(--font-mono);
-  margin-bottom: 0;
-  padding: 1rem;
-  flex-grow: 2;
-  align-self: center;
-}
-
-.navItem {
-  padding: 1rem;
-  border-inline-start: 2px solid var(--foreground);
-  font-family: var(--font-mono);
-  display: flex;
-  align-items: center;
-}
-
-.navItem a {
-  white-space: nowrap;
-}
-
 .footer {
   display: flex;
   flex-wrap: wrap-reverse;
@@ -74,7 +47,6 @@ p.homeLink {
 /* Tablet and Smaller Desktop */
 @media (min-width: 701px) {
   p.homeLink,
-  .navItem,
   .footerContent p,
   .footerItem {
     padding: 1.5rem 2rem;
