+++ app/(site)/components/header.module.css
@@ -1,0 +1,33 @@
+.nav {
+  display: flex;
+  justify-content: space-between;
+  align-items: stretch;
+  border-bottom: 2px solid var(--foreground);
+}
+
+p.homeLink {
+  font-family: var(--font-mono);
+  margin-bottom: 0;
+  padding: 1rem;
+  flex-grow: 2;
+  align-self: center;
+}
+
+.navItem {
+  padding: 1rem;
+  border-inline-start: 2px solid var(--foreground);
+  font-family: var(--font-mono);
+  display: flex;
+  align-items: center;
+}
+
+.navItem a {
+  white-space: nowrap;
+}
+
+/* Tablet and Smaller Desktop */
+@media (min-width: 701px) {
+  .navItem {
+    padding: 1.5rem 2rem;
+  }
+}
